Fix obvious compiler warnings: unused vars, sign-compare, catch-by-ref, snprintf size

+++ isis/src/base/objs/Spice/Spice.cpp
@@ -319,7 +319,7 @@ namespace Isis {
     }
     else {
       m_instrumentRotation->LoadCache(isd["instrument_pointing"]);
-      if (m_instrumentRotation->cacheSize() == m_instrumentRotation->GetFullCacheTime().size() && 
+      if ((size_t) m_instrumentRotation->cacheSize() == m_instrumentRotation->GetFullCacheTime().size() &&
           m_instrumentRotation->cacheSize() > 5) {
         m_instrumentRotation->MinimizeCache(SpiceRotation::DownsizeStatus::Yes);
         m_instrumentRotation->LoadTimeCache();
